requirements.txt
actions==1.0.3
allure-pytest==2.8.6
allure-python-commons==2.8.6
Appium-Python-Client==0.48
atomicwrites==1.3.0
attrs==19.3.0
bunch==1.0.1
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
docopt==0.6.2
enum34==1.1.6
idna==2.8
importlib-metadata==1.3.0
keyboard==0.13.4
lxml==4.4.2
more-itertools==8.0.2
namedlist==1.7
numpy==1.18.1
outcome==1.0.1
packaging==19.2
pandas==0.25.3
path==13.2.0
Pillow==7.0.0
pluggy==0.13.1
py==1.8.0
pyparsing==2.4.5
pytest==5.3.2
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
robotframework==3.1.2
ru==0.0.0
selenium==3.141.0
six==1.13.0
urllib3==1.25.7
wcwidth==0.1.7
yarg==0.1.9
zipp==0.6.0
